[✏️ RENAME] 불필요한 import문 제거 및 패키지 변경
